% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{create_custom_dataset}
\alias{create_custom_dataset}
\title{Create custom dataset}
\usage{
create_custom_dataset(
  area_type,
  variables,
  population_type = "UR",
  remove_zeroes = FALSE,
  column_names = "id"
)
}
\arguments{
\item{remove_zeroes}{}
}
\description{
Create custom dataset
}
