% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_available_area_types}
\alias{get_available_area_types}
\title{Get available area types}
\usage{
get_available_area_types(population_type = "UR")
}
\arguments{
\item{population_type}{}
}
\description{
Get available area types
}
