% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_available_areas}
\alias{get_available_areas}
\title{Get available areas}
\usage{
get_available_areas(area_type_id, search_param = NULL, population_type = "UR")
}
\arguments{
\item{population_type}{}
}
\description{
Get available areas
}
