% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_available_categorisations}
\alias{get_available_categorisations}
\title{Get available classifications}
\usage{
get_available_categorisations(variable, population_type = "UR", nest = FALSE)
}
\arguments{
\item{nest}{}
}
\description{
Get available classifications
}
