% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_available_variables}
\alias{get_available_variables}
\title{Get available variables}
\usage{
get_available_variables(search_param = NULL, population_type = "UR")
}
\arguments{
\item{population_type}{}
}
\description{
Get available variables
}
